+++ examples/for-tests-react-16/src/App.js
@@ -23,6 +23,7 @@ import DarkTheme from "./Themes/Dark";
 import HeliumTheme from "./Themes/Helium";
 import HydrogenTheme from "./Themes/Hydrogen";
 import { logWithPrefix } from "./logWithPrefix";
+import { ErrorBoundary } from "./ErrorBoundary";
 let { useNavigate } = require("react-router-dom");
 let withRouter = undefined;
 const loadv5RRD = window.localStorage.getItem("react-router-dom-is-v5") === "true";
@@ -222,6 +223,7 @@ let recipeList = [
 
 const testContext = {
     disableDefaultUI: getQueryParams("disableDefaultUI") === "true",
+    thirdPartyRedirectURL: localStorage.getItem("thirdPartyRedirectURL"),
 };
 
 if (authRecipe === "thirdparty") {
@@ -261,13 +263,25 @@ SuperTokens.init({
 /* App */
 function App() {
     if (doNotUseReactRouterDom) {
-        return <AppWithoutRouter />;
+        return (
+            <ErrorBoundary>
+                <AppWithoutRouter />
+            </ErrorBoundary>
+        );
     }
 
     if (loadv5RRD) {
-        return <AppWithReactDomRouterV5 authRecipe={authRecipe} />;
+        return (
+            <ErrorBoundary>
+                <AppWithReactDomRouterV5 authRecipe={authRecipe} />{" "}
+            </ErrorBoundary>
+        );
     } else {
-        return <AppWithReactDomRouter authRecipe={authRecipe} />;
+        return (
+            <ErrorBoundary>
+                <AppWithReactDomRouter authRecipe={authRecipe} />{" "}
+            </ErrorBoundary>
+        );
     }
 }
 
@@ -549,7 +563,7 @@ function getEmailPasswordConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
+function getThirdPartyPasswordlessConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyPasswordless.init({
         style: {
             container: {
@@ -688,6 +702,7 @@ function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
             defaultCountry: passwordlessDefaultCountry,
@@ -787,7 +802,7 @@ function getPasswordlessConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyConfigs({ disableDefaultUI }) {
+function getThirdPartyConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdParty.init({
         style: {
             container: {
@@ -881,6 +896,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },
@@ -891,7 +907,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
+function getThirdPartyEmailPasswordConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyEmailPassword.init({
         preAPIHook: async (context) => {
             console.log(`ST_LOGS THIRD_PARTY_EMAIL_PASSWORD PRE_API_HOOKS ${context.action}`);
@@ -1077,6 +1093,7 @@ function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },