+++ examples/for-tests/src/App.js
@@ -1,6 +1,7 @@
-import React, { Fragment, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import "./App.css";
 
+import { ErrorBoundary } from "./ErrorBoundary";
 import AppWithoutRouter from "./AppWithoutRouter";
 import AppWithReactDomRouter from "./AppWithReactDomRouter";
 import Footer from "./Footer";
@@ -233,6 +234,7 @@ let recipeList = [
 
 const testContext = {
     disableDefaultUI: getQueryParams("disableDefaultUI") === "true",
+    thirdPartyRedirectURL: localStorage.getItem("thirdPartyRedirectURL"),
 };
 
 if (authRecipe === "thirdparty") {
@@ -278,10 +280,18 @@ SuperTokens.init({
 /* App */
 function App() {
     if (doNotUseReactRouterDom) {
-        return <AppWithoutRouter />;
+        return (
+            <ErrorBoundary>
+                <AppWithoutRouter />
+            </ErrorBoundary>
+        );
     }
 
-    return <AppWithReactDomRouter authRecipe={authRecipe} />;
+    return (
+        <ErrorBoundary>
+            <AppWithReactDomRouter authRecipe={authRecipe} />
+        </ErrorBoundary>
+    );
 }
 
 function getQueryParams(param) {
@@ -632,7 +642,7 @@ function getEmailPasswordConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
+function getThirdPartyPasswordlessConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyPasswordless.init({
         style: {
             container: {
@@ -773,6 +783,7 @@ function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
             defaultCountry: passwordlessDefaultCountry,
@@ -881,7 +892,7 @@ function getPasswordlessConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyConfigs({ disableDefaultUI }) {
+function getThirdPartyConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdParty.init({
         style: {
             container: {
@@ -984,6 +995,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },
@@ -994,7 +1006,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
+function getThirdPartyEmailPasswordConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyEmailPassword.init({
         preAPIHook: async (context) => {
             if (localStorage.getItem(`SHOW_GENERAL_ERROR`) === `THIRD_PARTY_EMAIL_PASSWORD ${context.action}`) {
@@ -1189,6 +1201,7 @@ function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },