+++ examples/with-emailpassword/api-server.js
@@ -8,6 +8,7 @@ let { verifySession } = require("supertokens-node/recipe/session/framework/expre
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let EmailVerification = require("supertokens-node/recipe/emailverification");
 let EmailPassword = require("supertokens-node/recipe/emailpassword");
+let Dashboard = require("supertokens-node/recipe/dashboard");
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
 const apiDomain = process.env.REACT_APP_API_URL || `http://localhost:${apiPort}`;
@@ -26,7 +27,14 @@ supertokens.init({
         apiDomain, // TODO: Change to your app's API domain
         websiteDomain, // TODO: Change to your app's website domain
     },
-    recipeList: [EmailVerification.init({ mode: "REQUIRED" }), EmailPassword.init(), Session.init()],
+    recipeList: [
+        EmailVerification.init({ mode: "REQUIRED" }),
+        EmailPassword.init(),
+        Session.init(),
+        Dashboard.init({
+            apiKey: "supertokens_is_awesome",
+        }),
+    ],
 });
 
 const app = express();