+++ examples/with-thirdparty/api-server.js
@@ -9,6 +9,7 @@ let EmailVerification = require("supertokens-node/recipe/emailverification");
 let { verifySession } = require("supertokens-node/recipe/session/framework/express");
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let ThirdParty = require("supertokens-node/recipe/thirdparty");
+let Dashboard = require("supertokens-node/recipe/dashboard");
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
 const apiDomain = process.env.REACT_APP_API_URL || `http://localhost:${apiPort}`;
@@ -57,6 +58,9 @@ supertokens.init({
             },
         }),
         Session.init(),
+        Dashboard.init({
+            apiKey: "supertokens_is_awesome",
+        }),
     ],
 });