+++ examples/with-thirdpartypasswordless/api-server/index.js
@@ -8,6 +8,7 @@ let { verifySession } = require("supertokens-node/recipe/session/framework/expre
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let ThirdPartyPasswordless = require("supertokens-node/recipe/thirdpartypasswordless");
 let EmailVerification = require("supertokens-node/recipe/emailverification");
+let Dashboard = require("supertokens-node/recipe/dashboard");
 require("dotenv").config();
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
@@ -58,6 +59,9 @@ supertokens.init({
             flowType: "USER_INPUT_CODE_AND_MAGIC_LINK",
         }),
         Session.init(),
+        Dashboard.init({
+            apiKey: "supertokens_is_awesome",
+        }),
     ],
 });