+++ test/with-typescript/src/App.tsx
@@ -1334,10 +1334,6 @@ ThirdPartyPasswordless.consumePasswordlessCode({
         preAPIHook: undefined,
     },
 });
-// @ts-expect-error
-ThirdPartyPasswordless.consumePasswordlessCode(undefined);
-// @ts-expect-error
-ThirdPartyPasswordless.consumePasswordlessCode();
 
 ThirdPartyPasswordless.createPasswordlessCode({
     email: "",